Renamed the Resoruce folder to Resources to allow dynamic loading of resources at runtime. Added tags to allow sounds and images to be displayed with each event in the game. to add a sound include a flag in the format of {s:soudnName:volume}  example {s:spaceSound:0.5} to add an image include a flag in the format of {i:imageName}
Example with everything

[Earth]{visitedEarth, s:Space:0.5, i:spacefield}
Are you ready to takeoff to find the Gold of the Galaxy?
<>Yes[GoldYes]
<>No[GoldNo]
;

This event give the player the visitedEarth tag, which can be checked for later on if desired.
It also says that this event should play the sound Space at half volume and show the spacefield image.